requirements.txt
earthengine_api==0.1.324
ee==0.4
Fiona==1.8.18
kneebow==1.0.1
loess==2.1.2
numpy==1.19.5
pandas==1.1.5
progress==1.5
PyYAML==6.0
rasterio==1.1.8
scikit_learn==1.1.3
scipy==1.5.4
Shapely==1.7.1
